+++ poi/src/main/java/org/apache/poi/hssf/record/cont/ContinuableRecordInput.java
@@ -101,8 +101,8 @@ public class ContinuableRecordInput implements LittleEndianInput {
                 ((long)b5 << 40) +
                 ((long)b4 << 32) +
                 ((long)b3 << 24) +
-                (b2 << 16) +
-                (b1 <<  8) +
+                ((long)b2 << 16) +
+                ((long)b1 <<  8) +
                 (b0));
     }
